backend/requirements.txt
google-genai>=0.3.0
pydantic>=2.0.0
pydantic-ai>=0.0.1
openai>=2.0.0
anthropic>=0.7.0
opencv-python-headless
mediapipe
numpy
pandas
Pillow
httpx
aiohttp
uvicorn
fastapi
python-dotenv
python-multipart
pytz
sqlmodel
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-slugify
